// ocss-frontend/types/productsetservice.ts
export interface components {
  schemas: {
    ProductSetRequestDTO: {
      name?: string
      tenant?: string
      query?: string
      productIds?: string[]
    }
    ProductSetResponseDTO: {
      /** Format: int64 */
      id?: number
      name?: string
      tenant?: string
      query?: string
      productIds?: string[]
    }
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}